@@ -1,4 +1,4 @@
-class Athlete < ActiveRecord::Migration
+class Athlete < ActiveRecord::Migration[4.2]
   def change
     add_column :athletes, :username, :string
   end
