@@ -1,4 +1,4 @@
-class CreateActivities < ActiveRecord::Migration
+class CreateActivities < ActiveRecord::Migration[4.2]
   def change
     create_table :activities do |t|
       t.string :name
