@@ -1,4 +1,4 @@
-class RemoveActivityFromAthletes < ActiveRecord::Migration
+class RemoveActivityFromAthletes < ActiveRecord::Migration[4.2]
   def change
     remove_column :athletes, :activity, :string
   end
