@@ -1,4 +1,4 @@
-class AddAthleteToActivities < ActiveRecord::Migration
+class AddAthleteToActivities < ActiveRecord::Migration[4.2]
   def change
     add_reference :activities, :athlete, index: true
     add_foreign_key :activities, :athletes
