@@ -1,4 +1,4 @@
-class ChangeTypeColumnName < ActiveRecord::Migration
+class ChangeTypeColumnName < ActiveRecord::Migration[4.2]
   def change
     rename_column :activities, :type, :activity_type
   end
