@@ -1,4 +1,4 @@
-class AddAthleteNameAndAthleteIdAndImgUrlAndBeersAndCaloriesAndActivityIdsToLeaderboards < ActiveRecord::Migration
+class AddAthleteNameAndAthleteIdAndImgUrlAndBeersAndCaloriesAndActivityIdsToLeaderboards < ActiveRecord::Migration[4.2]
   def change
     add_column :leaderboards, :athlete_name, :string
     add_column :leaderboards, :img_url, :string
