@@ -1,4 +1,4 @@
-class AddActivitiesRefToLeaderboards < ActiveRecord::Migration
+class AddActivitiesRefToLeaderboards < ActiveRecord::Migration[4.2]
   def change
     add_reference :leaderboards, :activity, index: true
     add_foreign_key :leaderboards, :activities
