@@ -1,4 +1,4 @@
-class AddAthleteReferencesToLeaderboard < ActiveRecord::Migration
+class AddAthleteReferencesToLeaderboard < ActiveRecord::Migration[4.2]
   def change
     add_reference :leaderboards, :athlete, index: true
     add_foreign_key :leaderboards, :athletes
