@@ -1,4 +1,4 @@
-class AddLeaderboardToActivities < ActiveRecord::Migration
+class AddLeaderboardToActivities < ActiveRecord::Migration[4.2]
   def change
     add_reference :activities, :leaderboard, index: true
     add_foreign_key :activities, :leaderboards
